+++ js/exporter/jspdf/common/rows_spliting_utils/get_multipage_row_pages.js
@@ -27,10 +27,12 @@ const patchRects = (
         rect.sourceCellInfo.text = remainPageRects[rectIndex].sourceCellInfo.text;
         rect.h = remainPageRects[rectIndex].h;
     });
+
     const untouchedRowIdx = rectsToSplit.indexOf(rectsToPatch[rectsToPatch.length - 1]) + 1;
     if(untouchedRowIdx >= rectsToSplit.length) {
         return;
     }
+
     const delta = rectsToSplit[untouchedRowIdx].y - (rectsToPatch[0].y + remainPageRects[0].h);
     for(let idx = untouchedRowIdx; idx < rectsToSplit.length; idx++) {
         rectsToSplit[idx].y = rectsToSplit[idx].y - delta;
@@ -50,17 +52,21 @@ export const getMultiPageRowPages = (
     if(!splitMultiPageRowFunc) {
         return [];
     }
+
     const currentPageLastRect = currentPageRects[currentPageRects.length - 1];
     const nextPageFirstRect = rectsToSplit[currentPageRects.length];
     if(!nextPageFirstRect || isHeader(nextPageFirstRect)) {
         return [];
     }
+
     const isRectsFitsToPage = checkIsFitToPageFunc(isCurrentPageContainsOnlyHeader, nextPageFirstRect.h);
     if(isRectsFitsToPage && !isCurrentPageContainsOnlyHeader) {
         return [];
     }
+
     const rectsToPatch = rectsToSplit.filter(({ y }) => y === nextPageFirstRect.y);
     const firstRectYAdjustment = currentPageLastRect.y + currentPageLastRect.h;
+
     const [multiPageRowPages, remainPageRects] = spitMultiPageRows(
         rectsToPatch,
         isCurrentPageContainsOnlyHeader,
@@ -68,6 +74,8 @@ export const getMultiPageRowPages = (
         splitMultiPageRowFunc,
         checkIsFitToPageFunc
     );
+
     patchRects(rectsToSplit, rectsToPatch, remainPageRects);
+
     return multiPageRowPages;
 };
